flytt deltakere for kostnadsfordeling ut av UtbetalingService
For å gjøre ansvarsområdet til UtbetalingService litt tydligere
