+++ mulighetsrommet-api/src/main/kotlin/no/nav/mulighetsrommet/api/utbetaling/UtbetalingService.kt
@@ -1,7 +1,9 @@
 package no.nav.mulighetsrommet.api.utbetaling
 
-import arrow.core.*
-import io.ktor.http.*
+import arrow.core.Either
+import arrow.core.getOrElse
+import arrow.core.left
+import arrow.core.right
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.encodeToJsonElement
 import kotliquery.TransactionalSession
@@ -10,16 +12,9 @@ import no.nav.common.kafka.producer.feilhandtering.StoredProducerRecord
 import no.nav.mulighetsrommet.api.ApiDatabase
 import no.nav.mulighetsrommet.api.QueryContext
 import no.nav.mulighetsrommet.api.clients.kontoregisterOrganisasjon.KontoregisterOrganisasjonClient
-import no.nav.mulighetsrommet.api.clients.norg2.Norg2Client
-import no.nav.mulighetsrommet.api.clients.norg2.NorgError
-import no.nav.mulighetsrommet.api.clients.pdl.GeografiskTilknytning
-import no.nav.mulighetsrommet.api.clients.pdl.PdlGradering
-import no.nav.mulighetsrommet.api.clients.pdl.PdlIdent
-import no.nav.mulighetsrommet.api.clients.pdl.tilPersonNavn
 import no.nav.mulighetsrommet.api.endringshistorikk.DocumentClass
 import no.nav.mulighetsrommet.api.gjennomforing.model.GjennomforingDto
 import no.nav.mulighetsrommet.api.navansatt.model.Rolle
-import no.nav.mulighetsrommet.api.navenhet.db.NavEnhetDbo
 import no.nav.mulighetsrommet.api.responses.FieldError
 import no.nav.mulighetsrommet.api.responses.StatusResponse
 import no.nav.mulighetsrommet.api.responses.ValidationError
@@ -38,13 +33,9 @@ import no.nav.mulighetsrommet.api.utbetaling.api.UtbetalingType
 import no.nav.mulighetsrommet.api.utbetaling.db.DelutbetalingDbo
 import no.nav.mulighetsrommet.api.utbetaling.db.UtbetalingDbo
 import no.nav.mulighetsrommet.api.utbetaling.model.*
-import no.nav.mulighetsrommet.api.utbetaling.pdl.HentAdressebeskyttetPersonMedGeografiskTilknytningBolkPdlQuery
-import no.nav.mulighetsrommet.api.utbetaling.pdl.HentPersonBolkResponse
 import no.nav.mulighetsrommet.api.utbetaling.task.JournalforUtbetaling
 import no.nav.mulighetsrommet.ktor.exception.NotFound
-import no.nav.mulighetsrommet.ktor.exception.StatusException
 import no.nav.mulighetsrommet.model.*
-import no.nav.mulighetsrommet.tokenprovider.AccessType
 import no.nav.tiltak.okonomi.OkonomiBestillingMelding
 import no.nav.tiltak.okonomi.OpprettFaktura
 import no.nav.tiltak.okonomi.Tilskuddstype
@@ -63,8 +54,6 @@ class UtbetalingService(
     private val tilsagnService: TilsagnService,
     private val journalforUtbetaling: JournalforUtbetaling,
     private val kontoregisterOrganisasjonClient: KontoregisterOrganisasjonClient,
-    private val pdlQuery: HentAdressebeskyttetPersonMedGeografiskTilknytningBolkPdlQuery,
-    private val norg2Client: Norg2Client,
 ) {
     data class Config(
         val bestillingTopic: String,
@@ -96,84 +85,6 @@ class UtbetalingService(
                 logEndring("Utbetaling opprettet", dto, Tiltaksadministrasjon)
                 dto
             }
-    }
-
-    private suspend fun getPersoner(deltakerIdenter: List<NorskIdent>): Map<PdlIdent, Pair<HentPersonBolkResponse.Person, GeografiskTilknytning?>> {
-        val identer = deltakerIdenter
-            .map { ident -> PdlIdent(ident.value) }
-            .toNonEmptySetOrNull()
-            ?: return mapOf()
-
-        return pdlQuery.hentPersonOgGeografiskTilknytningBolk(identer, AccessType.M2M).getOrElse {
-            throw StatusException(
-                status = HttpStatusCode.InternalServerError,
-                detail = "Klarte ikke hente informasjon om personer og geografisk tilknytning",
-            )
-        }
-    }
-
-    private suspend fun hentEnhetForGeografiskTilknytning(geografiskTilknytning: GeografiskTilknytning): NavEnhetDbo? {
-        val norgEnhet = when (geografiskTilknytning) {
-            is GeografiskTilknytning.GtBydel -> norg2Client.hentEnhetByGeografiskOmraade(
-                geografiskTilknytning.value,
-            )
-
-            is GeografiskTilknytning.GtKommune -> norg2Client.hentEnhetByGeografiskOmraade(
-                geografiskTilknytning.value,
-            )
-
-            else -> return null
-        }
-
-        return norgEnhet
-            .map { hentNavEnhet(it.enhetNr) }
-            .getOrElse {
-                when (it) {
-                    NorgError.NotFound -> null
-                    NorgError.Error -> throw StatusException(
-                        HttpStatusCode.InternalServerError,
-                        "Fant ikke navenhet til geografisk tilknytning.",
-                    )
-                }
-            }
-    }
-
-    private fun hentNavEnhet(enhetsNummer: NavEnhetNummer) = db.session {
-        queries.enhet.get(enhetsNummer)
-    }
-
-    suspend fun getDeltakereForKostnadsfordeling(deltakerIdenter: List<NorskIdent>): Map<NorskIdent, DeltakerPerson> {
-        val personer = getPersoner(deltakerIdenter)
-
-        val deltakereForKostnadsfordeling = personer.map { (ident, pair) ->
-            val (person, geografiskTilknytning) = pair
-            val gradering = person.adressebeskyttelse.firstOrNull()?.gradering ?: PdlGradering.UGRADERT
-
-            if (gradering == PdlGradering.UGRADERT) {
-                val navEnhet = geografiskTilknytning?.let { hentEnhetForGeografiskTilknytning(it) }
-                val region = navEnhet?.overordnetEnhet?.let { hentNavEnhet(it) }
-                val navn = if (person.navn.isNotEmpty()) tilPersonNavn(person.navn) else "Ukjent"
-                val foedselsdato = if (person.foedselsdato.isNotEmpty()) person.foedselsdato.first() else null
-
-                DeltakerPerson(
-                    norskIdent = NorskIdent(ident.value),
-                    navn = navn,
-                    foedselsdato = foedselsdato?.foedselsdato,
-                    geografiskEnhet = navEnhet,
-                    region = region,
-                )
-            } else {
-                DeltakerPerson(
-                    norskIdent = NorskIdent(ident.value),
-                    navn = "Adressebeskyttet person",
-                    foedselsdato = null,
-                    geografiskEnhet = null,
-                    region = null,
-                )
-            }
-        }.associateBy { it.norskIdent }
-
-        return deltakereForKostnadsfordeling
     }
 
     suspend fun oppdaterUtbetalingBeregningForGjennomforing(id: UUID): Unit = db.transaction {
